+++ example/lib/common/tu_chong_source.dart
@@ -1,4 +1,5 @@
 import 'dart:convert' show json;
+import 'dart:math';
 
 import 'package:flutter/material.dart';
 
@@ -80,11 +81,14 @@ class TuChongItem {
   List<dynamic> rewardListPrefix;
   List<dynamic> sites;
   List<String> tags;
+  List<Color> tagColors = List<Color>();
   Site site;
 
   bool get hasImage {
     return images != null && images.length > 0;
   }
+
+  Size imageRawSize;
 
   Size get imageSize {
     if (!hasImage) return Size(0, 0);
@@ -93,11 +97,21 @@ class TuChongItem {
 
   String get imageUrl {
     if (!hasImage) return "";
-    return "https://photo.tuchong.com/" +
-        images[0].userId.toString() +
-        "/f/" +
-        images[0].imgId.toString() +
-        ".jpg";
+    return "https://photo.tuchong.com/${images[0].userId}/f/${images[0].imgId}.jpg";
+  }
+
+  String get avatarUrl => site.icon;
+
+  String get imageTitle {
+    if (!hasImage) return title;
+
+    return images[0].title;
+  }
+
+  String get imageDescription {
+    if (!hasImage) return content;
+
+    return images[0].description;
   }
 
   TuChongItem.fromParams(
@@ -208,9 +222,12 @@ class TuChongItem {
     }
 
     tags = jsonRes['tags'] == null ? null : [];
-
+    final int maxNum = 6;
     for (var tagsItem in tags == null ? [] : jsonRes['tags']) {
       tags.add(tagsItem);
+      tagColors.add(Color.fromARGB(255, Random.secure().nextInt(255),
+          Random.secure().nextInt(255), Random.secure().nextInt(255)));
+      if (tags.length == maxNum) break;
     }
 
     site = jsonRes['site'] == null ? null : new Site.fromJson(jsonRes['site']);